refactor: 공지사항 데이터 타입 변경으로 코드 수정
- 공지사항 테스트코드 같이 수정

+++ src/main/java/com/ahi/timecapsule/repository/NoticeRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 
   // 최신 날짜 기준 게시글 나오는 메서드
   Page<Notice> findAllByOrderByCreatedAtDesc(Pageable pageable);
